requirements-dev.txt
-r requirements.txt

black==24.1.1
cfgv==3.4.0
distlib==0.4.0
filelock==3.19.1
identify==2.6.15
iniconfig==2.1.0
mypy==1.8.0
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==4.3.0
pytest==7.4.0
pytest-flask==1.2.0
pyyaml==6.0.3
ruff==0.1.9
typing-extensions==4.15.0
virtualenv==20.34.0
